backend/requirements.txt
asgiref==3.2.10
django==2.2.19
django-filter~=21.1
djangorestframework~=3.12.4
pip==23.0.1
psycopg2-binary==2.8.6
python-dotenv==1.0.0
pytz==2023.3
setuptools==60.10.0
sqlparse==0.4.3
djoser==2.1.0
django-colorfield==0.8.0
gunicorn==20.1.0
django-cors-headers==2.2.0